Move test to MiscTypeTests
