+++ test/Npgsql.Tests/Types/MiscTypeTests.cs
@@ -3,6 +3,7 @@ using System.Data;
 using System.Threading.Tasks;
 using NpgsqlTypes;
 using NUnit.Framework;
+using NUnit.Framework.Constraints;
 
 namespace Npgsql.Tests.Types;
 
@@ -11,6 +12,9 @@ namespace Npgsql.Tests.Types;
 /// </summary>
 class MiscTypeTests : MultiplexingTestBase
 {
+    NpgsqlDataSource _slimWithoutMappingsDataSource = null!;
+    NpgsqlDataSource _slimWithMappingsDataSource = null!;
+
     [Test]
     public async Task Boolean()
     {
@@ -131,6 +135,30 @@ class MiscTypeTests : MultiplexingTestBase
     public Task Write_Record_is_not_supported()
         => AssertTypeUnsupportedWrite<object[], NotSupportedException>(new object[] { 1, "foo" }, "record");
 
+    [Test]
+    public async Task Records_supported_only_with_EnableRecords([Values] bool withMappings)
+    {
+        const string unsupportedMessage =
+            "Records aren't supported; please call EnableRecord on NpgsqlSlimDataSourceBuilder to enable records.";
+        Func<IResolveConstraint> assertExpr = () => withMappings
+            ? Throws.Nothing
+            : Throws.Exception
+                .TypeOf<NotSupportedException>()
+                .With.Property("Message").EqualTo(unsupportedMessage);
+
+        var dataSource = withMappings ? _slimWithMappingsDataSource : _slimWithoutMappingsDataSource;
+        await using var conn = await dataSource.OpenConnectionAsync();
+        await using var cmd = conn.CreateCommand();
+
+        // RecordHandler doesn't support writing, so we only check for reading
+        cmd.CommandText = "SELECT ('one'::text, 2)";
+        await using var reader = await cmd.ExecuteReaderAsync();
+        await reader.ReadAsync();
+
+        Assert.That(() => reader.GetValue(0), assertExpr());
+        Assert.That(() => reader.GetFieldValue<object[]>(0), assertExpr());
+    }
+
     #endregion Record
 
     [Test, Description("Makes sure that setting DbType.Object makes Npgsql infer the type")]
@@ -232,6 +260,23 @@ class MiscTypeTests : MultiplexingTestBase
         await using var cmd = new NpgsqlCommand("SELECT @p", conn);
         Assert.That(() => cmd.Parameters.Add(new NpgsqlParameter("p", DbType.UInt32) { Value = 8u }),
             Throws.Exception.TypeOf<NotSupportedException>());
+    }
+
+    [OneTimeSetUp]
+    public void SetUp()
+    {
+        _slimWithoutMappingsDataSource = new NpgsqlSlimDataSourceBuilder(ConnectionString)
+            .Build();
+        _slimWithMappingsDataSource = new NpgsqlSlimDataSourceBuilder(ConnectionString)
+            .EnableRecords()
+            .Build();
+    }
+
+    [OneTimeTearDown]
+    public void TearDown()
+    {
+        _slimWithoutMappingsDataSource.Dispose();
+        _slimWithMappingsDataSource.Dispose();
     }
 
     public MiscTypeTests(MultiplexingMode multiplexingMode) : base(multiplexingMode) {}
